feature: add vessel info panel
This commit creates the logic for tracking vessel under pointer.
Just using mouse events does not work very well, because the vessel
markers move and may disappear, so instead we keep track of latest
pointer position and when either mouse moves or a marker is
rendered we check the element under the cursor.

Text elements and the parent svg have pointer-events none so that
only vessel markers are taken into consideration.